pyproject.toml
[project]
name = "final"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11,<3.14"
dependencies = [
    "asyncpg>=0.29.0",
    "authlib>=1.3.0",
    "fastapi>=0.128.3",
    "celery>=5.3.6",
    "httpx>=0.28.1",
    "minio>=7.2.3",
    "openai>=2.17.0",
    "pydantic>=2.12.5",
    "pydantic-settings>=2.1.0",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.6",
    "python-dotenv>=1.2.1",
    "pyyaml>=6.0.1",
    "redis>=5.0.1",
    "uvicorn[standard]>=0.40.0",
    "email-validator>=2.3.0",
    "torch>=2.1.0",
    "torchaudio>=2.1.0",
    "transformers>=4.37.0",
    "kiwipiepy>=0.16.3",
    "librosa>=0.10.1",
    "soundfile>=0.12.1",
    "scikit-learn>=1.4.0",
    "scipy>=1.12.0",
    "numpy>=1.26.0",
    "psycopg2-binary>=2.9.9",
    "faster-whisper>=0.10.0",
    "SimpleITK>=2.3.0",
]

[dependency-groups]
dev = [
    "pytest>=9.0.2",
]
